Practical task of lesson 6 completed
1. Register models and set up an administrative section in the project.
    Add a link to the administrative section in the main menu
of the site, available only to superusers.
    Register the news model and add full text search.
    Register the lesson model and add fields to display in the table,
sorting, splitting the list into pages, filters, and an additional
action for a group of objects.
    Register the user model.
    Install the logo in the administrative section as on the main site.
2. Based on Django tools (CBV, authorization), add the ability to create
a user group that will be able to create, edit and delete news
in the main interface of the site. Restrict access to different
controllers by checking the appropriate permissions for the news model.
Depending on permissions, add or remove buttons for available actions
in the web interface.
3. Add the ability to leave grades with a comment for a course with
a form submission in asynchronous mode. Mandatory condition:
	the user can leave only one comment for the course.