Practical task of lesson 8 completed
1. In the tutorial project, implement a paginator object
on the news page.
2. Add an internationalization system to the project.
Implement the main menu in two languages ​​with the ability to switch.
3. Add to the project:
        a. Main page accessibility test.
        b. Tests to validate CRUD operations through the authenticated
user of the news objects.
        c. Tests for unavailability of CRUD operations for an
unauthenticated user of news objects.
        d. Test of the pending task to check the sending of e-mail.
        e. A test that checks the availability of a page that uses
a caching system. Implement isolation of the called controller
from the external system for the cache through a mock object.
        f. Tests using the Selenium tool. Check the transition
on the button to another page. Analyze the color match given
by the CSS style. Perform verification with an authenticated user.